requirements.txt
pillow
numpy
openocr-python
onnxruntime
mss
psutil
PyWinCtl